autodialer-app/requirements.txt
flask 
twilio 
python-dotenv 
langchain 
langchain-core
langchain-google-genai 
google-generativeai
